+++ source/core/transform/color_neon.cpp
@@ -62,9 +62,9 @@ void cvt_rgb_to_ycbcr_irrev_neon(int32_t *sp0, int32_t *sp1, int32_t *sp2, uint3
   double fY, fCb, fCr;
   int32x4_t R, G, B;
   float32x4_t Y, Cb, Cr, fR, fG, fB;
-  float32x4_t a0 = vdupq_n_f32(ALPHA_R);
-  float32x4_t a1 = vdupq_n_f32(ALPHA_G);
-  float32x4_t a2 = vdupq_n_f32(ALPHA_B);
+  float32x4_t a0 = vdupq_n_f32(static_cast<float32_t>(ALPHA_R));
+  float32x4_t a1 = vdupq_n_f32(static_cast<float32_t>(ALPHA_G));
+  float32x4_t a2 = vdupq_n_f32(static_cast<float32_t>(ALPHA_B));
   // for (uint32_t n = 0; n < num_tc_samples - num_tc_samples % 4; n += 4) {
   for (; num_tc_samples > 4; num_tc_samples -= 4) {
     R  = vld1q_s32(sp0);
@@ -77,8 +77,8 @@ void cvt_rgb_to_ycbcr_irrev_neon(int32_t *sp0, int32_t *sp1, int32_t *sp2, uint3
     Y  = vmlaq_f32(Y, fG, a1);
     Y  = vmlaq_f32(Y, fB, a2);
     // Y  = fR * a0 + fG * a1 + fB * a2;
-    Cb = vmulq_n_f32(vsubq_f32(fB, Y), 1.0 / CB_FACT_B);
-    Cr = vmulq_n_f32(vsubq_f32(fR, Y), 1.0 / CR_FACT_R);
+    Cb = vmulq_n_f32(vsubq_f32(fB, Y), static_cast<float32_t>(1.0 / CB_FACT_B));
+    Cr = vmulq_n_f32(vsubq_f32(fR, Y), static_cast<float32_t>(1.0 / CR_FACT_R));
 
     // TODO: need to consider precision and setting FPSCR register value
     vst1q_s32(sp0, vcvtnq_s32_f32(Y));
@@ -144,10 +144,13 @@ void cvt_ycbcr_to_rgb_irrev_neon(int32_t *sp0, int32_t *sp1, int32_t *sp2, uint3
     float32x4_t f32x4_Cr = vcvtq_f32_s32(s32x4_Cr);
 
     // TODO: need to consider precision and setting FPSCR register value
-    vst1q_s32(sp0 + n, vcvtnq_s32_f32(vaddq_f32(f32x4_Y, vmulq_n_f32(f32x4_Cr, CR_FACT_R))));
-    vst1q_s32(sp2 + n, vcvtnq_s32_f32(vaddq_f32(f32x4_Y, vmulq_n_f32(f32x4_Cb, CB_FACT_B))));
-    vst1q_s32(sp1 + n, vcvtnq_s32_f32(vsubq_f32(f32x4_Y, vaddq_f32(vmulq_n_f32(f32x4_Cr, CR_FACT_G),
-                                                                   vmulq_n_f32(f32x4_Cb, CB_FACT_G)))));
+    vst1q_s32(sp0 + n,
+              vcvtnq_s32_f32(vaddq_f32(f32x4_Y, vmulq_n_f32(f32x4_Cr, static_cast<float32_t>(CR_FACT_R)))));
+    vst1q_s32(sp2 + n,
+              vcvtnq_s32_f32(vaddq_f32(f32x4_Y, vmulq_n_f32(f32x4_Cb, static_cast<float32_t>(CB_FACT_B)))));
+    vst1q_s32(sp1 + n, vcvtnq_s32_f32(vsubq_f32(
+                           f32x4_Y, vaddq_f32(vmulq_n_f32(f32x4_Cr, static_cast<float32_t>(CR_FACT_G)),
+                                              vmulq_n_f32(f32x4_Cb, static_cast<float32_t>(CB_FACT_G))))));
   }
   for (uint32_t n = num_tc_samples - num_tc_samples % 4; n < num_tc_samples; ++n) {
     fY     = static_cast<double>(sp0[n]);
